+++ duckscript_sdk/src/sdk/std/var/set_by_name/mod_test.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::sdk::std::var::set;
 use crate::test;
 use crate::test::CommandValidation;
 
@@ -14,11 +15,18 @@ fn run_no_arguments() {
 
 #[test]
 fn run_only_name() {
-    test::run_script_and_error(vec![create("")], "out = set_by_name test", "out");
+    test::run_script_and_validate(
+        vec![create(""), set::create("")],
+        r#"
+        out = set test
+        set_by_name out
+        "#,
+        CommandValidation::None,
+    );
 }
 
 #[test]
-fn run_valid() {
+fn run_name_and_value() {
     let context = test::run_script_and_validate(
         vec![create("")],
         "out = set_by_name test value",